ECA-1073: Better time resolution for CRL-properties. SimpleTime helper class can be reused for more user-friendly input of time intervals.
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/branches/Branch_3_9/ejbca@7762 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/test/org/ejbca/core/ejb/ca/crl/TestCreateCRLSession.java
@@ -303,7 +303,7 @@ public class TestCreateCRLSession extends TestCase {
 		log.trace(">test05CRLPeriodOverflow()");
 		// Fetch CAInfo and save CRLPeriod
 		CAInfo cainfo = TestTools.getCAAdminSession().getCAInfo(admin, caid);
-		int tempCRLPeriod = cainfo.getCRLPeriod();
+		long tempCRLPeriod = cainfo.getCRLPeriod();
 		try {
 			// Create a user that Should be revoked
 			boolean userExists = false;
